Allow checks to specify depdencies for checks on other clients.
This extends the current dependecy-filtering mechanism, which
looks for events for a given check on the same client, to optionally
look for a triggered event on a different client.  Specifying "foo"
will look for an event for the "foo" check on the same client,
whereas specifying "myclient2/foo" will check for an event for the
"foo" check on the "myclient2" client.

The use-case for this is having external checks, i.e. ping or http
checks for a service on clients that are not running the service.
The external accessibility checks supercede any host-level failures.