Fix bad error message if you use an unknown CLI option.
This causes the check's constructor to fail, and thus subsequently calling check.critical fails
as the constructor never returned a check object, this results in a second exception, meaning
that the error message about the check is never shown to the user.

Rather than calling check.cricial, we now do what it does manually, so that we'll always print the
unexpected exception that we caught.