More accurate refresh interval on first resend
Say I have a check where interval = 120, occurrences = 14, refresh = 1800 (the default). That means "number" is 1800.fdiv(120).to_i, which is 15.

At 14 occurrences, an alert is sent out.
At 15 occurrences, 15 % 15 is 0... so another alert is sent out.

In the PR, I subtract the occurrences threshold first before comparing to the refresh:
At 14 occurrences, an alert is sent out.
At 15 occurrences, (15 - 14) % 15 = 1. No alert is sent.
At 29 occurrences, (29 - 14) % 15 = 0. Another alert is sent, 30 minutes after the original, as was probably intended by the user.

This seems more in line with the idea of the refresh interval to me.