exit unknown on exception or bad args
when a plugin throws an unexpected exception we should exit(3) unknown.
As it means the plugin was unable to signal warning or critical itself
meaning the state of the thing being checked is unknown. This commit
changes the behavior from exit(2) critical which generally results in
someone getting paged with a false alarm.

Similarly, when someone deploys a check and has given it the wrong
arguments we should also exit(3) unknown. We don't know the state of the
thing we are attempting to check.  We only know the check was
misconfigured because the options are wrong.