introduce acknowledgments
short version:  if a stash has type: "acknowledgment" then on resolve handle the
event and delete the stash.

I propose people need two workflows with regards to silencing alerts.

1. downtime, the thing being checked may fail and resolve multiple times
during some time interval, alerts should be filtered.

2. acknowldgement, the thing being checked just failed, someone has
acknowledged this fact and future failing alerts should be filtered,
until the check is resolved.

right now sensu-handler only supports the notion of downtime with it's
silence stashes. this has a few problems related to an acknowledgment
workflow

- if the event is resolved during the downtime window the handler is
  prevented from handling it. for example say a pagerduty alert was
  created on event create but the resolve event was filtered
  it requires manually resolving the pager duty event.
- user must remember to remove the stash. or guess an appropriate expiration time.
  (in my observation our users never set one. and then forget them)
- if downtime is too long, someone may believe they fixed the issue only
  to have it come back and not alert again until the downtime window is
  closed.
- if downtime window is too short, we are needlessly alerted more when
  it expires.

This commit introduces acknowlegments by way of the type field in the
normal /silence/* stashes if the type field is set to 'acknowledgment'
and the event action is resolve then the stash will be deleted and the
event processed.